vm: extract all call/selfdestruct/log-related logic from EEI
